improve

+++ worker/fuzzer/src/RTC/FuzzerDtlsTransport.cpp
@@ -51,10 +51,10 @@ void FuzzerRtcDtlsTransport::Fuzz(const uint8_t* data, size_t len)
 		// NOTE: Use a random integer in range 1..5 since FingerprintAlgorithm enum
 		// has 5 possible values starting with value 1.
 		dtlsRemoteFingerprint.algorithm = static_cast<RTC::DtlsTransport::FingerprintAlgorithm>(
-		  Utils::Crypto::GetRandomUInt<uint16_t>(1u, 5u));
+		  Utils::Crypto::GetRandomUInt<uint8_t>(1u, 5u));
 
 		dtlsRemoteFingerprint.value =
-		  Utils::Crypto::GetRandomString(Utils::Crypto::GetRandomUInt<uint16_t>(3u, 20u));
+		  Utils::Crypto::GetRandomString(Utils::Crypto::GetRandomUInt<uint8_t>(3u, 20u));
 
 		dtlsTransportSingleton->Run(localRole);
 		dtlsTransportSingleton->SetRemoteFingerprint(dtlsRemoteFingerprint);
